Added in a simple death system and a very simple system to handle attacks. Will be testing these new systems soon with a very simple gui

+++ Dwarf/ECS/Messages/damage_entity_message.h
@@ -12,10 +12,10 @@ enum class DamageTypes
 struct damage_entity_message : public base_message_t
 {
 	damage_entity_message() = default;
-	damage_entity_message(double dmg, size_t attacker, size_t target) 
+	damage_entity_message(int dmg, size_t attacker, size_t target) 
 		: dmg(dmg), attacker(attacker), target(target) {}
 
-	double dmg = 0.0;
+	int dmg = 0;
 
 	size_t attacker = 0;
 	size_t target = 0;
